Add initial support for service auto-scaling to ECS service.
This change adds service auto-scaling support for tasks in ECS.
ECS is getting complicated enough that I've done some refactoring
into separate modules for some larger concerns in the deployer.

I would like to start testing some of these pieces in a bit more
fine detail now that ECS is getting more complex, but I haven't
added any additional tests in this PR yet. The current tests are
passing.
