+++ lib/aws/ssm-calls.js
@@ -49,4 +49,13 @@ exports.deleteParameters = function (parameterNames) {
         .then(() => {
             return true;
         })
+        .catch(err => {
+            //If the param is already gone we won't count it as an error
+            if(err.code === "ParameterNotFound") { 
+                return true;
+            }
+            else {
+                throw err;
+            }
+        });
 }
