+++ lib/lifecycles/check.js
@@ -38,9 +38,7 @@ exports.check = function (handelFile) {
             'subnet-eeeeeeee',
             'subnet-99999999'
         ],
-        ecs_ami: 'ami-66666666',
-        ssh_bastion_sg: 'sg-44444444',
-        on_prem_cidr: '10.10.10.10/0'
+        ssh_bastion_sg: 'sg-44444444'
     }).getAccountConfig();
 
     //Load all the currently implemented service deployers from the 'services' directory
